requirements.txt
asgiref==3.3.4
Django==3.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
django-cors-headers==3.7.0
